issue-083 안쓰는 필드, 뷰 삭제
