+++ app/src/main/AndroidManifest.xml
@@ -31,8 +31,6 @@
         <activity android:name=".presentation.upload.UploadActivity"
             android:windowSoftInputMode="stateAlwaysHidden"/>
 
-        <activity android:name=".presentation.main.FrameActivity" />
-
     </application>
 
 </manifest>
