--- app/src/main/java/com/boostcamp/dreampicker/presentation/feed/list/FeedListViewModel.java
@@ -1,6 +1,0 @@
-package com.boostcamp.dreampicker.presentation.feed.list;
-
-import com.boostcamp.dreampicker.presentation.BaseViewModel;
-
-public class FeedListViewModel extends BaseViewModel {
-}
