--- app/src/main/java/com/boostcamp/dreampicker/presentation/user/list/UserListViewModel.java
@@ -1,6 +1,0 @@
-package com.boostcamp.dreampicker.presentation.user.list;
-
-import com.boostcamp.dreampicker.presentation.BaseViewModel;
-
-public class UserListViewModel extends BaseViewModel {
-}
