--- app/src/main/res/layout/fragment_user_list.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycler_user_list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"/>
-</layout>
