--- app/src/main/res/layout/item_feed_list.xml
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-        <variable
-            name="feed"
-            type="com.boostcamp.dreampicker.data.model.Feed"/>
-    </data>
-
-    <androidx.cardview.widget.CardView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:cardCornerRadius="5dp"
-        app:cardElevation="4dp"
-        app:cardUseCompatPadding="true">
-
-        <FrameLayout
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/feed_photo_image_height">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:orientation="horizontal">
-
-                <ImageView
-                    android:id="@+id/iv_left"
-                    android:layout_width="0dp"
-                    android:layout_weight="1"
-                    android:layout_height="match_parent"
-                    style="@style/ImageView"
-                    android:scaleType="centerCrop"
-                    app:image="@{feed.imageList.get(0).image}"/>
-
-                <ImageView
-                    android:id="@+id/iv_right"
-                    android:layout_width="0dp"
-                    android:layout_weight="1"
-                    android:layout_height="match_parent"
-                    style="@style/ImageView"
-                    android:scaleType="centerCrop"
-                    app:image="@{feed.imageList.get(1).image}"/>
-            </LinearLayout>
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:orientation="vertical"
-                android:background="@color/colorBlackTransParent">
-
-                <LinearLayout
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:orientation="horizontal"
-                    android:gravity="center_vertical">
-
-                    <ImageView
-                        android:layout_width="@dimen/profile_image_size_median"
-                        android:layout_height="@dimen/profile_image_size_median"
-                        android:layout_margin="@dimen/space_median"
-                        app:circleImage="@{feed.user.profileImageResource}"/>
-
-                    <TextView
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:textSize="@dimen/text_small"
-                        android:textColor="@android:color/white"
-                        android:text="@{feed.user.name}"/>
-                </LinearLayout>
-
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:layout_margin="@dimen/space_median"
-                    android:gravity="bottom"
-                    android:textColor="@android:color/white"
-                    android:textSize="@dimen/text_s_large"
-                    android:text="@{feed.content}"/>
-
-            </LinearLayout>
-        </FrameLayout>
-    </androidx.cardview.widget.CardView>
-
-</layout>
