--- app/src/main/res/layout/item_user_list.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <data>
-        <variable
-            name="user"
-            type="com.boostcamp.dreampicker.data.model.User"/>
-    </data>
-
-    <androidx.cardview.widget.CardView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:cardCornerRadius="5dp"
-        app:cardElevation="4dp"
-        app:cardUseCompatPadding="true">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:gravity="center_vertical"
-            android:orientation="horizontal">
-
-            <ImageView
-                android:layout_width="@dimen/profile_image_size_median"
-                android:layout_height="@dimen/profile_image_size_median"
-                android:layout_margin="@dimen/space_median"
-                style="@style/ImageView"
-                android:src="@drawable/profile"/>
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textSize="@dimen/text_large"
-                android:textColor="@color/colorBlack"
-                android:text="@{user.name}"/>
-
-        </LinearLayout>
-    </androidx.cardview.widget.CardView>
-</layout>
